services/airflow/requirements.txt
kafka-python
snowflake-connector-python
apache-airflow-providers-snowflake
